Rename operation module
The placeholder 'orginal' (for the orignal paper's operations) was
not very descriptive. These operations are all convolutional,
renaming this accordingly makes more sense.
